Crate verdant_parsers_gitdep

verdant-parsers-gitdep

Collection of tree-sitter parsers for verdant, using cargo git dependencies.

See the project overview for more information on all parser collections.

List of included parsers

List of parsers included in the some feature

bash

c

cpp

css

dockerfile (not supported by this collection)

gitattributes (not supported by this collection)

gitcommit (not supported by this collection)

gitconfig (not supported by this collection)

gitdiff (not supported by this collection)

gitignore (not supported by this collection)

gitrebase (not supported by this collection)

go

graphql (not supported by this collection)

html

java

javascript

json

kotlin

lua

proto (not supported by this collection)

python

rust

toml (not supported by this collection)

tsx

typescript

yaml

List of parsers additionally included in the most feature

asm

c_sharp

clojure (not supported by this collection)

cmake

comment (not supported by this collection)

csv (not supported by this collection)

dart (not supported by this collection)

diff

elixir

fish

haskell

hcl (not supported by this collection)

ini (not supported by this collection)

jsdoc

json5 (not supported by this collection)

jsonc (not supported by this collection)

luap (not supported by this collection)

make

markdown

markdown_inline

nginx (not supported by this collection)

nix

php

php_only

printf

ps1 (not supported by this collection)

regex

ruby

scala

scss (not supported by this collection)

sql

sshconfig (not supported by this collection)

swift (not supported by this collection)

terraform (not supported by this collection)

tsv (not supported by this collection)

typst (not supported by this collection)

xml (not supported by this collection)

zig

zsh (not supported by this collection)

List of parsers additionally included in the all feature

ada (not supported by this collection)

agda (not supported by this collection)

apex (not supported by this collection)

arduino (not supported by this collection)

asm68k (not supported by this collection)

astro (not supported by this collection)

authzed (not supported by this collection)

awk (not supported by this collection)

bass (not supported by this collection)

beancount (not supported by this collection)

bib (not supported by this collection)

bibtex (not supported by this collection)

bicep (not supported by this collection)

bitbake (not supported by this collection)

blade (not supported by this collection)

blueprint (not supported by this collection)

bp (not supported by this collection)

bpftrace (not supported by this collection)

brightscript (not supported by this collection)

bzl (not supported by this collection)

c3 (not supported by this collection)

caddy (not supported by this collection)

cairo (not supported by this collection)

capnp (not supported by this collection)

chatito (not supported by this collection)

circom (not supported by this collection)

clarity (not supported by this collection)

cook (not supported by this collection)

corn (not supported by this collection)

cpon (not supported by this collection)

csharp (not supported by this collection)

cuda (not supported by this collection)

cue (not supported by this collection)

cylc (not supported by this collection)

d (not supported by this collection)

desktop (not supported by this collection)

dhall (not supported by this collection)

disassembly (not supported by this collection)

djot (not supported by this collection)

dot (not supported by this collection)

doxygen (not supported by this collection)

dtd (not supported by this collection)

dts (not supported by this collection)

earthfile (not supported by this collection)

ebnf (not supported by this collection)

editorconfig (not supported by this collection)

eds (not supported by this collection)

eelixir (not supported by this collection)

ejs

elisp (not supported by this collection)

elm (not supported by this collection)

elsa (not supported by this collection)

elvish (not supported by this collection)

enforce (not supported by this collection)

erb

erlang (not supported by this collection)

eruby (not supported by this collection)

faust (not supported by this collection)

fennel (not supported by this collection)

fidl (not supported by this collection)

firrtl (not supported by this collection)

foam (not supported by this collection)

forth (not supported by this collection)

fortran (not supported by this collection)

fsd (not supported by this collection)

fsh (not supported by this collection)

fsharp (not supported by this collection)

func (not supported by this collection)

fusion (not supported by this collection)

gap (not supported by this collection)

gaptst (not supported by this collection)

gdresource (not supported by this collection)

gdscript (not supported by this collection)

gdshaderinc (not supported by this collection)

gleam

glimmer_javascript (not supported by this collection)

glimmer_typescript (not supported by this collection)

glsl (not supported by this collection)

gn (not supported by this collection)

gnuplot (not supported by this collection)

goctl (not supported by this collection)

gomod (not supported by this collection)

gosum (not supported by this collection)

gotmpl (not supported by this collection)

gowork (not supported by this collection)

gpg (not supported by this collection)

gren (not supported by this collection)

groovy (not supported by this collection)

groq (not supported by this collection)

gstlaunch (not supported by this collection)

hack (not supported by this collection)

handlebars (not supported by this collection)

hare (not supported by this collection)

haskellpersistent (not supported by this collection)

haxe (not supported by this collection)

heex (not supported by this collection)

helm (not supported by this collection)

help (not supported by this collection)

hexdump (not supported by this collection)

hjson (not supported by this collection)

hlsl (not supported by this collection)

hlsplaylist (not supported by this collection)

hocon (not supported by this collection)

hoon (not supported by this collection)

htmlangular (not supported by this collection)

htmldjango (not supported by this collection)

http (not supported by this collection)

hurl (not supported by this collection)

hyprlang (not supported by this collection)

idl (not supported by this collection)

idris2 (not supported by this collection)

inko (not supported by this collection)

ispc (not supported by this collection)

janet (not supported by this collection)

javadoc (not supported by this collection)

jinja (not supported by this collection)

jinja_inline (not supported by this collection)

jproperties (not supported by this collection)

jq (not supported by this collection)

jsonnet (not supported by this collection)

julia

just (not supported by this collection)

kcl (not supported by this collection)

kconfig (not supported by this collection)

kdl (not supported by this collection)

kitty (not supported by this collection)

kos (not supported by this collection)

koto (not supported by this collection)

kusto (not supported by this collection)

lalrpop

latex (not supported by this collection)

ld (not supported by this collection)

ledger (not supported by this collection)

leo (not supported by this collection)

liquid (not supported by this collection)

liquidsoap (not supported by this collection)

lisp (not supported by this collection)

llvm (not supported by this collection)

luadoc (not supported by this collection)

luau (not supported by this collection)

magik (not supported by this collection)

matlab (not supported by this collection)

menhir (not supported by this collection)

mermaid (not supported by this collection)

meson (not supported by this collection)

mlir (not supported by this collection)

muttrc (not supported by this collection)

nasm (not supported by this collection)

nickel (not supported by this collection)

nim (not supported by this collection)

nim_format_string (not supported by this collection)

ninja (not supported by this collection)

nqc (not supported by this collection)

nu (not supported by this collection)

objc (not supported by this collection)

objdump (not supported by this collection)

ocaml

ocaml_interface

ocamlinterface (not supported by this collection)

ocamllex (not supported by this collection)

odin (not supported by this collection)

pascal (not supported by this collection)

passwd (not supported by this collection)

pbtxt (not supported by this collection)

pem (not supported by this collection)

perl (not supported by this collection)

pgn (not supported by this collection)

phpdoc (not supported by this collection)

pioasm (not supported by this collection)

pkl (not supported by this collection)

po (not supported by this collection)

pod (not supported by this collection)

poefilter (not supported by this collection)

pony (not supported by this collection)

prisma (not supported by this collection)

problog (not supported by this collection)

prolog (not supported by this collection)

promql (not supported by this collection)

prql (not supported by this collection)

psv (not supported by this collection)

pug (not supported by this collection)

puppet (not supported by this collection)

purescript (not supported by this collection)

pymanifest (not supported by this collection)

ql

qml (not supported by this collection)

qmldir (not supported by this collection)

query (not supported by this collection)

r (not supported by this collection)

racket (not supported by this collection)

ralph (not supported by this collection)

rasi (not supported by this collection)

razor (not supported by this collection)

rbs (not supported by this collection)

re2c (not supported by this collection)

readline (not supported by this collection)

rego (not supported by this collection)

requirements (not supported by this collection)

rescript (not supported by this collection)

rifleconf (not supported by this collection)

rnoweb (not supported by this collection)

robot (not supported by this collection)

robots (not supported by this collection)

roc (not supported by this collection)

ron (not supported by this collection)

rst (not supported by this collection)

runescript (not supported by this collection)

rush

scfg (not supported by this collection)

scheme (not supported by this collection)

sface (not supported by this collection)

sflog (not supported by this collection)

shaderslang (not supported by this collection)

slim (not supported by this collection)

slint (not supported by this collection)

smali (not supported by this collection)

smithy (not supported by this collection)

snakemake (not supported by this collection)

solidity (not supported by this collection)

soql (not supported by this collection)

sosl (not supported by this collection)

sourcepawn (not supported by this collection)

sparql (not supported by this collection)

sproto (not supported by this collection)

squirrel (not supported by this collection)

strace (not supported by this collection)

styled (not supported by this collection)

supercollider (not supported by this collection)

superhtml (not supported by this collection)

svelte (not supported by this collection)

sway (not supported by this collection)

sxhkdrc (not supported by this collection)

systemtap (not supported by this collection)

tablegen (not supported by this collection)

tact (not supported by this collection)

tal (not supported by this collection)

tape (not supported by this collection)

tcl (not supported by this collection)

teal (not supported by this collection)

templ (not supported by this collection)

tera (not supported by this collection)

tex (not supported by this collection)

thrift (not supported by this collection)

tiger (not supported by this collection)

tla (not supported by this collection)

tmux (not supported by this collection)

todotxt (not supported by this collection)

trace32 (not supported by this collection)

turtle (not supported by this collection)

twig (not supported by this collection)

typespec (not supported by this collection)

typoscript (not supported by this collection)

udevrules (not supported by this collection)

ungrammar (not supported by this collection)

unison (not supported by this collection)

ursa (not supported by this collection)

usd (not supported by this collection)

v (not supported by this collection)

vala (not supported by this collection)

verilog (not supported by this collection)

vhdl (not supported by this collection)

vim (not supported by this collection)

vrl (not supported by this collection)

vto (not supported by this collection)

vue (not supported by this collection)

wat (not supported by this collection)

wgsl (not supported by this collection)

wgsl_bevy (not supported by this collection)

wing (not supported by this collection)

wit (not supported by this collection)

wxml (not supported by this collection)

xcompose (not supported by this collection)

xdefaults (not supported by this collection)

yang (not supported by this collection)

yuck (not supported by this collection)

zathurarc (not supported by this collection)

ziggy (not supported by this collection)

ziggy_schema (not supported by this collection)

Features

Every supported language has a feature with the same name as the respective public function. Additionally the three feature groups some, most, and all are available.

some — Include parsers for the most widely known supported languages.

most — Implies some. Include parsers for most common languages.

all — Implies most. Include parsers for all supported languages.

docs — Meant to be enabled when building docs

Structs

LanguageSetImpl

An implementation of LanguageSet including all languages in the enabled feature set.

Enums

Lang

An enum of every supported language in the current feature set.

Constants

LANGUAGES

A list of all languages supported by the current feature set.

LANGUAGE_NAMES

A list of all language names supported by the current feature set.

Functions

ada

This parser is not supported by this parser collection and thus this function will panic!

agda

This parser is not supported by this parser collection and thus this function will panic!

apex

This parser is not supported by this parser collection and thus this function will panic!

arduino

This parser is not supported by this parser collection and thus this function will panic!

asm

Get the parser for asm.

asm68k

This parser is not supported by this parser collection and thus this function will panic!

astro

This parser is not supported by this parser collection and thus this function will panic!

authzed

This parser is not supported by this parser collection and thus this function will panic!

awk

This parser is not supported by this parser collection and thus this function will panic!

bash

Get the parser for bash.

bass

This parser is not supported by this parser collection and thus this function will panic!

beancount

This parser is not supported by this parser collection and thus this function will panic!

bib

This parser is not supported by this parser collection and thus this function will panic!

bibtex

This parser is not supported by this parser collection and thus this function will panic!

bicep

This parser is not supported by this parser collection and thus this function will panic!

bitbake

This parser is not supported by this parser collection and thus this function will panic!

blade

This parser is not supported by this parser collection and thus this function will panic!

blueprint

This parser is not supported by this parser collection and thus this function will panic!

bp

This parser is not supported by this parser collection and thus this function will panic!

bpftrace

This parser is not supported by this parser collection and thus this function will panic!

brightscript

This parser is not supported by this parser collection and thus this function will panic!

bzl

This parser is not supported by this parser collection and thus this function will panic!

c

Get the parser for c.

c3

This parser is not supported by this parser collection and thus this function will panic!

c_sharp

Get the parser for c_sharp.

caddy

This parser is not supported by this parser collection and thus this function will panic!

cairo

This parser is not supported by this parser collection and thus this function will panic!

capnp

This parser is not supported by this parser collection and thus this function will panic!

chatito

This parser is not supported by this parser collection and thus this function will panic!

circom

This parser is not supported by this parser collection and thus this function will panic!

clarity

This parser is not supported by this parser collection and thus this function will panic!

clojure

This parser is not supported by this parser collection and thus this function will panic!

cmake

Get the parser for cmake.

comment

This parser is not supported by this parser collection and thus this function will panic!

cook

This parser is not supported by this parser collection and thus this function will panic!

corn

This parser is not supported by this parser collection and thus this function will panic!

cpon

This parser is not supported by this parser collection and thus this function will panic!

cpp

Get the parser for cpp.

csharp

This parser is not supported by this parser collection and thus this function will panic!

css

Get the parser for css.

csv

This parser is not supported by this parser collection and thus this function will panic!

cuda

This parser is not supported by this parser collection and thus this function will panic!

cue

This parser is not supported by this parser collection and thus this function will panic!

cylc

This parser is not supported by this parser collection and thus this function will panic!

d

This parser is not supported by this parser collection and thus this function will panic!

dart

This parser is not supported by this parser collection and thus this function will panic!

desktop

This parser is not supported by this parser collection and thus this function will panic!

dhall

This parser is not supported by this parser collection and thus this function will panic!

diff

Get the parser for diff.

disassembly

This parser is not supported by this parser collection and thus this function will panic!

djot

This parser is not supported by this parser collection and thus this function will panic!

dockerfile

This parser is not supported by this parser collection and thus this function will panic!

dot

This parser is not supported by this parser collection and thus this function will panic!

doxygen

This parser is not supported by this parser collection and thus this function will panic!

dtd

This parser is not supported by this parser collection and thus this function will panic!

dts

This parser is not supported by this parser collection and thus this function will panic!

earthfile

This parser is not supported by this parser collection and thus this function will panic!

ebnf

This parser is not supported by this parser collection and thus this function will panic!

editorconfig

This parser is not supported by this parser collection and thus this function will panic!

eds

This parser is not supported by this parser collection and thus this function will panic!

eelixir

This parser is not supported by this parser collection and thus this function will panic!

ejs

Get the parser for ejs.

elisp

This parser is not supported by this parser collection and thus this function will panic!

elixir

Get the parser for elixir.

elm

This parser is not supported by this parser collection and thus this function will panic!

elsa

This parser is not supported by this parser collection and thus this function will panic!

elvish

This parser is not supported by this parser collection and thus this function will panic!

enforce

This parser is not supported by this parser collection and thus this function will panic!

erb

Get the parser for erb.

erlang

This parser is not supported by this parser collection and thus this function will panic!

eruby

This parser is not supported by this parser collection and thus this function will panic!

faust

This parser is not supported by this parser collection and thus this function will panic!

fennel

This parser is not supported by this parser collection and thus this function will panic!

fidl

This parser is not supported by this parser collection and thus this function will panic!

firrtl

This parser is not supported by this parser collection and thus this function will panic!

fish

Get the parser for fish.

foam

This parser is not supported by this parser collection and thus this function will panic!

forth

This parser is not supported by this parser collection and thus this function will panic!

fortran

This parser is not supported by this parser collection and thus this function will panic!

fsd

This parser is not supported by this parser collection and thus this function will panic!

fsh

This parser is not supported by this parser collection and thus this function will panic!

fsharp

This parser is not supported by this parser collection and thus this function will panic!

func

This parser is not supported by this parser collection and thus this function will panic!

fusion

This parser is not supported by this parser collection and thus this function will panic!

gap

This parser is not supported by this parser collection and thus this function will panic!

gaptst

This parser is not supported by this parser collection and thus this function will panic!

gdresource

This parser is not supported by this parser collection and thus this function will panic!

gdscript

This parser is not supported by this parser collection and thus this function will panic!

gdshaderinc

This parser is not supported by this parser collection and thus this function will panic!

gitattributes

This parser is not supported by this parser collection and thus this function will panic!

gitcommit

This parser is not supported by this parser collection and thus this function will panic!

gitconfig

This parser is not supported by this parser collection and thus this function will panic!

gitdiff

This parser is not supported by this parser collection and thus this function will panic!

gitignore

This parser is not supported by this parser collection and thus this function will panic!

gitrebase

This parser is not supported by this parser collection and thus this function will panic!

gleam

Get the parser for gleam.

glimmer_javascript

This parser is not supported by this parser collection and thus this function will panic!

glimmer_typescript

This parser is not supported by this parser collection and thus this function will panic!

glsl

This parser is not supported by this parser collection and thus this function will panic!

gn

This parser is not supported by this parser collection and thus this function will panic!

gnuplot

This parser is not supported by this parser collection and thus this function will panic!

go

Get the parser for go.

goctl

This parser is not supported by this parser collection and thus this function will panic!

gomod

This parser is not supported by this parser collection and thus this function will panic!

gosum

This parser is not supported by this parser collection and thus this function will panic!

gotmpl

This parser is not supported by this parser collection and thus this function will panic!

gowork

This parser is not supported by this parser collection and thus this function will panic!

gpg

This parser is not supported by this parser collection and thus this function will panic!

graphql

This parser is not supported by this parser collection and thus this function will panic!

gren

This parser is not supported by this parser collection and thus this function will panic!

groovy

This parser is not supported by this parser collection and thus this function will panic!

groq

This parser is not supported by this parser collection and thus this function will panic!

gstlaunch

This parser is not supported by this parser collection and thus this function will panic!

hack

This parser is not supported by this parser collection and thus this function will panic!

handlebars

This parser is not supported by this parser collection and thus this function will panic!

hare

This parser is not supported by this parser collection and thus this function will panic!

haskell

Get the parser for haskell.

haskellpersistent

This parser is not supported by this parser collection and thus this function will panic!

haxe

This parser is not supported by this parser collection and thus this function will panic!

hcl

This parser is not supported by this parser collection and thus this function will panic!

heex

This parser is not supported by this parser collection and thus this function will panic!

helm

This parser is not supported by this parser collection and thus this function will panic!

help

This parser is not supported by this parser collection and thus this function will panic!

hexdump

This parser is not supported by this parser collection and thus this function will panic!

hjson

This parser is not supported by this parser collection and thus this function will panic!

hlsl

This parser is not supported by this parser collection and thus this function will panic!

hlsplaylist

This parser is not supported by this parser collection and thus this function will panic!

hocon

This parser is not supported by this parser collection and thus this function will panic!

hoon

This parser is not supported by this parser collection and thus this function will panic!

html

Get the parser for html.

htmlangular

This parser is not supported by this parser collection and thus this function will panic!

htmldjango

This parser is not supported by this parser collection and thus this function will panic!

http

This parser is not supported by this parser collection and thus this function will panic!

hurl

This parser is not supported by this parser collection and thus this function will panic!

hyprlang

This parser is not supported by this parser collection and thus this function will panic!

idl

This parser is not supported by this parser collection and thus this function will panic!

idris2

This parser is not supported by this parser collection and thus this function will panic!

ini

This parser is not supported by this parser collection and thus this function will panic!

inko

This parser is not supported by this parser collection and thus this function will panic!

ispc

This parser is not supported by this parser collection and thus this function will panic!

janet

This parser is not supported by this parser collection and thus this function will panic!

java

Get the parser for java.

javadoc

This parser is not supported by this parser collection and thus this function will panic!

javascript

Get the parser for javascript.

jinja

This parser is not supported by this parser collection and thus this function will panic!

jinja_inline

This parser is not supported by this parser collection and thus this function will panic!

jproperties

This parser is not supported by this parser collection and thus this function will panic!

jq

This parser is not supported by this parser collection and thus this function will panic!

jsdoc

Get the parser for jsdoc.

json

Get the parser for json.

json5

This parser is not supported by this parser collection and thus this function will panic!

jsonc

This parser is not supported by this parser collection and thus this function will panic!

jsonnet

This parser is not supported by this parser collection and thus this function will panic!

julia

Get the parser for julia.

just

This parser is not supported by this parser collection and thus this function will panic!

kcl

This parser is not supported by this parser collection and thus this function will panic!

kconfig

This parser is not supported by this parser collection and thus this function will panic!

kdl

This parser is not supported by this parser collection and thus this function will panic!

kitty

This parser is not supported by this parser collection and thus this function will panic!

kos

This parser is not supported by this parser collection and thus this function will panic!

kotlin

Get the parser for kotlin.

koto

This parser is not supported by this parser collection and thus this function will panic!

kusto

This parser is not supported by this parser collection and thus this function will panic!

lalrpop

Get the parser for lalrpop.

latex

This parser is not supported by this parser collection and thus this function will panic!

ld

This parser is not supported by this parser collection and thus this function will panic!

ledger

This parser is not supported by this parser collection and thus this function will panic!

leo

This parser is not supported by this parser collection and thus this function will panic!

liquid

This parser is not supported by this parser collection and thus this function will panic!

liquidsoap

This parser is not supported by this parser collection and thus this function will panic!

lisp

This parser is not supported by this parser collection and thus this function will panic!

llvm

This parser is not supported by this parser collection and thus this function will panic!

lua

Get the parser for lua.

luadoc

This parser is not supported by this parser collection and thus this function will panic!

luap

This parser is not supported by this parser collection and thus this function will panic!

luau

This parser is not supported by this parser collection and thus this function will panic!

magik

This parser is not supported by this parser collection and thus this function will panic!

make

Get the parser for make.

markdown

Get the parser for markdown.

markdown_inline

Get the parser for markdown_inline.

matlab

This parser is not supported by this parser collection and thus this function will panic!

menhir

This parser is not supported by this parser collection and thus this function will panic!

mermaid

This parser is not supported by this parser collection and thus this function will panic!

meson

This parser is not supported by this parser collection and thus this function will panic!

mlir

This parser is not supported by this parser collection and thus this function will panic!

muttrc

This parser is not supported by this parser collection and thus this function will panic!

nasm

This parser is not supported by this parser collection and thus this function will panic!

nginx

This parser is not supported by this parser collection and thus this function will panic!

nickel

This parser is not supported by this parser collection and thus this function will panic!

nim

This parser is not supported by this parser collection and thus this function will panic!

nim_format_string

This parser is not supported by this parser collection and thus this function will panic!

ninja

This parser is not supported by this parser collection and thus this function will panic!

nix

Get the parser for nix.

nqc

This parser is not supported by this parser collection and thus this function will panic!

nu

This parser is not supported by this parser collection and thus this function will panic!

objc

This parser is not supported by this parser collection and thus this function will panic!

objdump

This parser is not supported by this parser collection and thus this function will panic!

ocaml

Get the parser for ocaml.

ocaml_interface

Get the parser for ocaml_interface.

ocamlinterface

This parser is not supported by this parser collection and thus this function will panic!

ocamllex

This parser is not supported by this parser collection and thus this function will panic!

odin

This parser is not supported by this parser collection and thus this function will panic!

pascal

This parser is not supported by this parser collection and thus this function will panic!

passwd

This parser is not supported by this parser collection and thus this function will panic!

pbtxt

This parser is not supported by this parser collection and thus this function will panic!

pem

This parser is not supported by this parser collection and thus this function will panic!

perl

This parser is not supported by this parser collection and thus this function will panic!

pgn

This parser is not supported by this parser collection and thus this function will panic!

php

Get the parser for php.

php_only

Get the parser for php_only.

phpdoc

This parser is not supported by this parser collection and thus this function will panic!

pioasm

This parser is not supported by this parser collection and thus this function will panic!

pkl

This parser is not supported by this parser collection and thus this function will panic!

po

This parser is not supported by this parser collection and thus this function will panic!

pod

This parser is not supported by this parser collection and thus this function will panic!

poefilter

This parser is not supported by this parser collection and thus this function will panic!

pony

This parser is not supported by this parser collection and thus this function will panic!

printf

Get the parser for printf.

prisma

This parser is not supported by this parser collection and thus this function will panic!

problog

This parser is not supported by this parser collection and thus this function will panic!

prolog

This parser is not supported by this parser collection and thus this function will panic!

promql

This parser is not supported by this parser collection and thus this function will panic!

proto

This parser is not supported by this parser collection and thus this function will panic!

prql

This parser is not supported by this parser collection and thus this function will panic!

ps1

This parser is not supported by this parser collection and thus this function will panic!

psv

This parser is not supported by this parser collection and thus this function will panic!

pug

This parser is not supported by this parser collection and thus this function will panic!

puppet

This parser is not supported by this parser collection and thus this function will panic!

purescript

This parser is not supported by this parser collection and thus this function will panic!

pymanifest

This parser is not supported by this parser collection and thus this function will panic!

python

Get the parser for python.

ql

Get the parser for ql.

qml

This parser is not supported by this parser collection and thus this function will panic!

qmldir

This parser is not supported by this parser collection and thus this function will panic!

query

This parser is not supported by this parser collection and thus this function will panic!

r

This parser is not supported by this parser collection and thus this function will panic!

racket

This parser is not supported by this parser collection and thus this function will panic!

ralph

This parser is not supported by this parser collection and thus this function will panic!

rasi

This parser is not supported by this parser collection and thus this function will panic!

razor

This parser is not supported by this parser collection and thus this function will panic!

rbs

This parser is not supported by this parser collection and thus this function will panic!

re2c

This parser is not supported by this parser collection and thus this function will panic!

readline

This parser is not supported by this parser collection and thus this function will panic!

regex

Get the parser for regex.

rego

This parser is not supported by this parser collection and thus this function will panic!

requirements

This parser is not supported by this parser collection and thus this function will panic!

rescript

This parser is not supported by this parser collection and thus this function will panic!

rifleconf

This parser is not supported by this parser collection and thus this function will panic!

rnoweb

This parser is not supported by this parser collection and thus this function will panic!

robot

This parser is not supported by this parser collection and thus this function will panic!

robots

This parser is not supported by this parser collection and thus this function will panic!

roc

This parser is not supported by this parser collection and thus this function will panic!

ron

This parser is not supported by this parser collection and thus this function will panic!

rst

This parser is not supported by this parser collection and thus this function will panic!

ruby

Get the parser for ruby.

runescript

This parser is not supported by this parser collection and thus this function will panic!

rush

Get the parser for rush.

rust

Get the parser for rust.

scala

Get the parser for scala.

scfg

This parser is not supported by this parser collection and thus this function will panic!

scheme

This parser is not supported by this parser collection and thus this function will panic!

scss

This parser is not supported by this parser collection and thus this function will panic!

sface

This parser is not supported by this parser collection and thus this function will panic!

sflog

This parser is not supported by this parser collection and thus this function will panic!

shaderslang

This parser is not supported by this parser collection and thus this function will panic!

slim

This parser is not supported by this parser collection and thus this function will panic!

slint

This parser is not supported by this parser collection and thus this function will panic!

smali

This parser is not supported by this parser collection and thus this function will panic!

smithy

This parser is not supported by this parser collection and thus this function will panic!

snakemake

This parser is not supported by this parser collection and thus this function will panic!

solidity

This parser is not supported by this parser collection and thus this function will panic!

soql

This parser is not supported by this parser collection and thus this function will panic!

sosl

This parser is not supported by this parser collection and thus this function will panic!

sourcepawn

This parser is not supported by this parser collection and thus this function will panic!

sparql

This parser is not supported by this parser collection and thus this function will panic!

sproto

This parser is not supported by this parser collection and thus this function will panic!

sql

Get the parser for sql.

squirrel

This parser is not supported by this parser collection and thus this function will panic!

sshconfig

This parser is not supported by this parser collection and thus this function will panic!

strace

This parser is not supported by this parser collection and thus this function will panic!

styled

This parser is not supported by this parser collection and thus this function will panic!

supercollider

This parser is not supported by this parser collection and thus this function will panic!

superhtml

This parser is not supported by this parser collection and thus this function will panic!

svelte

This parser is not supported by this parser collection and thus this function will panic!

sway

This parser is not supported by this parser collection and thus this function will panic!

swift

This parser is not supported by this parser collection and thus this function will panic!

sxhkdrc

This parser is not supported by this parser collection and thus this function will panic!

systemtap

This parser is not supported by this parser collection and thus this function will panic!

tablegen

This parser is not supported by this parser collection and thus this function will panic!

tact

This parser is not supported by this parser collection and thus this function will panic!

tal

This parser is not supported by this parser collection and thus this function will panic!

tape

This parser is not supported by this parser collection and thus this function will panic!

tcl

This parser is not supported by this parser collection and thus this function will panic!

teal

This parser is not supported by this parser collection and thus this function will panic!

templ

This parser is not supported by this parser collection and thus this function will panic!

tera

This parser is not supported by this parser collection and thus this function will panic!

terraform

This parser is not supported by this parser collection and thus this function will panic!

tex

This parser is not supported by this parser collection and thus this function will panic!

thrift

This parser is not supported by this parser collection and thus this function will panic!

tiger

This parser is not supported by this parser collection and thus this function will panic!

tla

This parser is not supported by this parser collection and thus this function will panic!

tmux

This parser is not supported by this parser collection and thus this function will panic!

todotxt

This parser is not supported by this parser collection and thus this function will panic!

toml

This parser is not supported by this parser collection and thus this function will panic!

trace32

This parser is not supported by this parser collection and thus this function will panic!

tsv

This parser is not supported by this parser collection and thus this function will panic!

tsx

Get the parser for tsx.

turtle

This parser is not supported by this parser collection and thus this function will panic!

twig

This parser is not supported by this parser collection and thus this function will panic!

typescript

Get the parser for typescript.

typespec

This parser is not supported by this parser collection and thus this function will panic!

typoscript

This parser is not supported by this parser collection and thus this function will panic!

typst

This parser is not supported by this parser collection and thus this function will panic!

udevrules

This parser is not supported by this parser collection and thus this function will panic!

ungrammar

This parser is not supported by this parser collection and thus this function will panic!

unison

This parser is not supported by this parser collection and thus this function will panic!

ursa

This parser is not supported by this parser collection and thus this function will panic!

usd

This parser is not supported by this parser collection and thus this function will panic!

v

This parser is not supported by this parser collection and thus this function will panic!

vala

This parser is not supported by this parser collection and thus this function will panic!

verilog

This parser is not supported by this parser collection and thus this function will panic!

vhdl

This parser is not supported by this parser collection and thus this function will panic!

vim

This parser is not supported by this parser collection and thus this function will panic!

vrl

This parser is not supported by this parser collection and thus this function will panic!

vto

This parser is not supported by this parser collection and thus this function will panic!

vue

This parser is not supported by this parser collection and thus this function will panic!

wat

This parser is not supported by this parser collection and thus this function will panic!

wgsl

This parser is not supported by this parser collection and thus this function will panic!

wgsl_bevy

This parser is not supported by this parser collection and thus this function will panic!

wing

This parser is not supported by this parser collection and thus this function will panic!

wit

This parser is not supported by this parser collection and thus this function will panic!

wxml

This parser is not supported by this parser collection and thus this function will panic!

xcompose

This parser is not supported by this parser collection and thus this function will panic!

xdefaults

This parser is not supported by this parser collection and thus this function will panic!

xml

This parser is not supported by this parser collection and thus this function will panic!

yaml

Get the parser for yaml.

yang

This parser is not supported by this parser collection and thus this function will panic!

yuck

This parser is not supported by this parser collection and thus this function will panic!

zathurarc

This parser is not supported by this parser collection and thus this function will panic!

zig

Get the parser for zig.

ziggy

This parser is not supported by this parser collection and thus this function will panic!

ziggy_schema

This parser is not supported by this parser collection and thus this function will panic!

zsh

This parser is not supported by this parser collection and thus this function will panic!